Split auth configure options into class
