+++ osu.Server.Spectator/Startup.cs
@@ -1,22 +1,15 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
-using System.IdentityModel.Tokens.Jwt;
-using System.IO;
-using System.Security.Cryptography;
-using Dapper;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Microsoft.IdentityModel.Tokens;
+using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
-using Org.BouncyCastle.Crypto.Parameters;
-using Org.BouncyCastle.Security;
-using osu.Server.Spectator.Database;
+using osu.Server.Spectator.Authentication;
 using osu.Server.Spectator.Hubs;
 
 namespace osu.Server.Spectator
@@ -49,46 +42,13 @@ namespace osu.Server.Spectator
 
         protected virtual void ConfigureAuthentication(IServiceCollection services)
         {
-            var rsa = getKeyProvider();
-
+            services.AddSingleton<IConfigureOptions<JwtBearerOptions>, ConfigureJwtBearerOptions>();
             services.AddAuthentication(config =>
-            {
-                config.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                config.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            }).AddJwtBearer(options =>
-            {
-                options.TokenValidationParameters = new TokenValidationParameters
-                {
-                    IssuerSigningKey = new RsaSecurityKey(rsa),
-                    // TODO: check what "5" means.
-                    ValidAudience = "5",
-                    // TODO: figure out why this isn't included in the token.
-                    ValidateIssuer = false,
-                    ValidIssuer = "https://osu.ppy.sh/"
-                };
-
-                options.Events = new JwtBearerEvents
-                {
-                    OnTokenValidated = async context =>
                     {
-                        var jwtToken = (JwtSecurityToken)context.SecurityToken;
-                        int tokenUserId = int.Parse(jwtToken.Subject);
-
-                        using (var conn = DB.GetConnection())
-                        {
-                            // check expiry/revocation against database
-                            var userId = await conn.QueryFirstOrDefaultAsync<int?>("SELECT user_id FROM oauth_access_tokens WHERE revoked = false AND expires_at > now() AND id = @id",
-                                new { id = jwtToken.Id });
-
-                            if (userId != tokenUserId)
-                            {
-                                Console.WriteLine("Token revoked or expired");
-                                context.Fail("Token has expired or been revoked");
-                            }
-                        }
-                    },
-                };
-            });
+                        config.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+                        config.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+                    })
+                    .AddJwtBearer(null); // options will be injected through DI, via the singleton registration above.
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -108,33 +68,6 @@ namespace osu.Server.Spectator
 
             app.UseEndpoints(endpoints => { endpoints.MapHub<SpectatorHub>("/spectator"); });
             app.UseEndpoints(endpoints => { endpoints.MapHub<MultiplayerHub>("/multiplayer"); });
-        }
-
-        /// <summary>
-        /// borrowed from https://stackoverflow.com/a/54323524
-        /// </summary>
-        private static RSACryptoServiceProvider getKeyProvider()
-        {
-            string key = File.ReadAllText("oauth-public.key");
-
-            key = key.Replace("-----BEGIN PUBLIC KEY-----", "");
-            key = key.Replace("-----END PUBLIC KEY-----", "");
-            key = key.Replace("\n", "");
-
-            var keyBytes = Convert.FromBase64String(key);
-
-            var asymmetricKeyParameter = PublicKeyFactory.CreateKey(keyBytes);
-            var rsaKeyParameters = (RsaKeyParameters)asymmetricKeyParameter;
-            var rsaParameters = new RSAParameters
-            {
-                Modulus = rsaKeyParameters.Modulus.ToByteArrayUnsigned(),
-                Exponent = rsaKeyParameters.Exponent.ToByteArrayUnsigned()
-            };
-
-            var rsa = new RSACryptoServiceProvider();
-            rsa.ImportParameters(rsaParameters);
-
-            return rsa;
         }
     }
 }
